Removed regularization loss on the final layer, which improved accuracy and loss of the model.

+++ NumberDetectorGPU/Loss.cuh
@@ -5,16 +5,16 @@
 class Loss
 {
 public:
-	Matrix<half>* m_predictions;
-	Matrix<half>* m_ground_truth;
-	Matrix<half>* m_one_hot_encoded_ground_truth;
+	Matrix<float>* m_predictions;
+	Matrix<float>* m_ground_truth;
+	Matrix<float>* m_one_hot_encoded_ground_truth;
 
 	// Loss
-	Matrix<half>* m_negative_log_confidencies;
+	Matrix<float>* m_negative_log_confidencies;
 	float m_data_loss;
 
 	// Accuracy
-	Matrix<half>* m_num_predictions_equal_ground_truth;
+	Matrix<float>* m_num_predictions_equal_ground_truth;
 	float m_data_accuracy;
 
 	Loss()
@@ -32,11 +32,11 @@ public:
 		m_data_accuracy = m_num_predictions_equal_ground_truth->ColMean();
 	}
 
-	half* GetPredictions()
+	float* GetPredictions()
 	{
-		half* predictions = new half[3];
+		float* predictions = new float[3];
 
-		cudaMemcpy(predictions, m_predictions->GetMatrix(), 3 * sizeof(half), cudaMemcpyDeviceToHost);
+		cudaMemcpy(predictions, m_predictions->GetMatrix(), 3 * sizeof(float), cudaMemcpyDeviceToHost);
 
 		return predictions;
 	}
@@ -51,7 +51,7 @@ public:
 		return m_data_accuracy;
 	}
 
-	Matrix<half>* GetDinputs()
+	Matrix<float>* GetDinputs()
 	{
 		return m_dinputs;
 	}
@@ -64,7 +64,7 @@ protected:
 	
 
 	// Backpropagation
-	Matrix<half>* m_dinputs;
+	Matrix<float>* m_dinputs;
 };
 
 template <class T>
